chore: rename layer to wavs in bash scripts
