+++ operator/rust/crates/operator/src/register_wavs_operator.rs
